+++ dsh_sdk/src/dsh/bootstrap.rs
@@ -142,10 +142,7 @@ pub(crate) enum DshCall<'a> {
     /// Call to retreive datastreams.json.
     Datastream(&'a DshConfig),
     /// Call to post the certificate signing request.
-    CertificateSignRequest {
-        config: &'a DshConfig,
-        csr: picky::pem::Pem<'a>,
-    },
+    CertificateSignRequest { config: &'a DshConfig, csr: &'a str },
 }
 
 impl DshCall<'_> {
@@ -291,10 +288,10 @@ mod tests {
             DshCall::Datastream(&dsh_config).request_builder("https://test_host", &Client::new());
         let request = builder.build().unwrap();
         assert_eq!(request.method().as_str(), "GET");
-        let pem = picky::pem::Pem::new("test_type", "test".as_bytes());
+        let csr = "-----BEGIN test_type-----\n-----END test_type-----";
         let builder: reqwest::blocking::RequestBuilder = DshCall::CertificateSignRequest {
             config: &dsh_config,
-            csr: pem,
+            csr,
         }
         .request_builder("https://test_host", &Client::new());
         let request = builder.build().unwrap();
@@ -309,7 +306,7 @@ mod tests {
             "test_token"
         );
         let body = from_utf8(request.body().unwrap().as_bytes().unwrap()).unwrap();
-        assert!(body.contains("-----BEGIN test_type-----"));
+        assert_eq!(body, csr);
     }
 
     #[test]